backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
markdown-it-py==3.0.0

